refactor: update the location of the PromissoryNotes section
