+++ src/pages/board/outlets/boardlayout/PromissoryNotes/config.tsx
