+++ src/pages/board/outlets/boardlayout/PromissoryNotes/index.tsx
