+++ src/routes/board.tsx
@@ -6,7 +6,7 @@ import { BoardLayout } from "@pages/board/outlets/boardlayout";
 import { FinancialReporting } from "@pages/board/outlets/financialReporting";
 import { Requirements } from "@pages/board/outlets/boardlayout/Requirements";
 import { dataRequirements } from "@pages/board/outlets/boardlayout/Requirements/config";
-import { PromissoryNotes } from "@pages/board/outlets/PromissoryNotes";
+import { PromissoryNotes } from "@pages/board/outlets/boardlayout/PromissoryNotes";
 
 function BoardRoutes() {
   return (
